feat: Add solution for Robot Return to Origin
This commit introduces a solution to the "Robot Return to Origin" problem.
The `judgeCircle` method simulates the movement of a robot on a 2D plane and determines if it returns to the origin (0, 0) after a sequence of moves.
The moves are represented by a string containing the characters U (up), D (down), L (left), and R (right).