feat: Add solution for Time Based Key-Value Store (LeetCode 981)
This commit adds the solution for the LeetCode problem 981 "Time Based Key-Value Store". The solution uses a HashMap to store key-value pairs and a list of pairs (timestamp, value) for each key. Binary search is used to efficiently retrieve the value corresponding to the largest timestamp less than or equal to the given timestamp. The file also includes a detailed explanation of the monotonic property that makes binary search applicable here.