feat: Add Minimum Consecutive Cards to Pick Up solution
Adds a solution for LeetCode 2260 "Minimum Consecutive Cards to Pick Up" using a HashMap.

The file `SlidingWindow/MinimumConsecutiveCardsToPickUp.java` implements a solution to find the minimum number of consecutive cards you have to pick up to have a pair of matching cards.

Also, adds a comment to the LongestSubstringWithoutRepeatingCharacters solution to clarify the logic.